restored libaed2

+++ build_glm.sh
@@ -24,12 +24,13 @@ while [ $# -gt 0 ] ; do
   shift
 done
 
+export MAKE=make
 export OSTYPE=`uname -s`
 if [ "$OSTYPE" = "Darwin" ] ; then
   if [ "$HOMEBREW" = "" ] ; then
-    brew -v >& /dev/null
+    brew -v > /dev/null 2>&1
     if [ $? != 0 ] ; then
-      which port >& /dev/null
+      which port > /dev/null 2>&1
       if [ $? != 0 ] ; then
         echo no ports and no brew
       else
@@ -39,40 +40,47 @@ if [ "$OSTYPE" = "Darwin" ] ; then
       export HOMEBREW=true
     fi
   fi
+else
+  if [ "$OSTYPE" = "FreeBSD" ] ; then
+    export FC=flang
+    export MAKE=gmake
+  fi
 fi
 
 
-# if FC is not defined we look for gfortran-8 first because some systems
-# will have gfortran at version 7 but also gfortran version 8 as gfortran-8
-# if we can't find gfortran default to ifort
+# see if FC is defined, if not look for gfortran at least v8
 if [ "$FC" = "" ] ; then
-  gfortran-8 -v > /dev/null 2>&1
+  gfortran -v > /dev/null 2>&1
   if [ $? != 0 ] ; then
-    gfortran -v > /dev/null 2>&1
-    if [ $? != 0 ] ; then
-      export FC=ifort
+    export FC=ifort
+  else
+    VERS=`gfortran -dumpversion | cut -d\. -f1`
+    if [ $VERS -ge 8 ] ; then
+      export FC=gfortran
     else
-      VERS=`gfortran -dumpversion | cut -d\. -f1`
-      if [ $VERS -ge 8 ] ; then
-        export FC=gfortran
-      else
+      gfortran-8 -v > /dev/null 2>&1
+      if [ $? != 0 ] ; then
         export FC=ifort
+      else
+        export gfortran-8
       fi
     fi
-  else
-    export FC=gfortran-8
   fi
 fi
 
 if [ "$FC" = "ifort" ] ; then
-   if [ -d /opt/intel/bin ] ; then
-      . /opt/intel/bin/compilervars.sh intel64
-   fi
-   which ifort > /dev/null 2>&1
-   if [ $? != 0 ] ; then
-      echo ifort compiler requested, but not found
-      exit 1
-   fi
+  if [ -d /opt/intel/oneapi ] ; then
+     . /opt/intel/oneapi/setvars.sh
+  else
+    if [ -d /opt/intel/bin ] ; then
+       . /opt/intel/bin/compilervars.sh intel64
+    fi
+    which ifort > /dev/null 2>&1
+    if [ $? != 0 ] ; then
+       echo ifort compiler requested, but not found
+       exit 1
+    fi
+  fi
 fi
 
 export F77=$FC
@@ -81,6 +89,9 @@ export F95=$FC
 
 export MPI=OPENMPI
 
+if [ "$AED2DIR" = "" ] ; then
+  export AED2DIR=../libaed2
+fi
 if [ "$PLOTDIR" = "" ] ; then
   export PLOTDIR=../libplot
 fi
@@ -116,56 +127,76 @@ if [ "$FABM" = "true" ] ; then
   else
     cmake ${FABMDIR}/src || exit 1
   fi
-  make || exit 1
+  ${MAKE} || exit 1
+fi
+
+if [ "${AED2}" = "true" ] ; then
+  cd ${AED2DIR}
+  ${MAKE} || exit 1
+  cd ..
+  if [ "${AED2PLS}" != "" ] ; then
+    if [ -d ${AED2PLS} ] ; then
+      cd ${AED2PLS}
+      ${MAKE} || exit 1
+      cd ..
+    fi
+  fi
 fi
 
 if [ "${AED}" = "true" ] ; then
   cd  ${CURDIR}/../libaed-water
-  make || exit 1
+  ${MAKE} || exit 1
   DAEDWATDIR=`pwd`
   if [ -d ${CURDIR}/../libaed-benthic ] ; then
     echo build libaed-benthic
     cd  ${CURDIR}/../libaed-benthic
-    make || exit 1
+    ${MAKE} || exit 1
     DAEDBENDIR=`pwd`
   fi
   if [ -d ${CURDIR}/../libaed-riparian ] ; then
     echo build libaed-riparian
     cd  ${CURDIR}/../libaed-riparian
-    make || exit 1
+    ${MAKE} || exit 1
     DAEDRIPDIR=`pwd`
   fi
   if [ -d ${CURDIR}/../libaed-demo ] ; then
     echo build libaed-demo
     cd  ${CURDIR}/../libaed-demo
-    make || exit 1
+    ${MAKE} || exit 1
     DAEDDMODIR=`pwd`
   fi
   if [ -d ${CURDIR}/../libaed-dev ] ; then
     echo build libaed-dev
     cd  ${CURDIR}/../libaed-dev
-    make || exit 1
+    ${MAKE} || exit 1
     DAEDDEVDIR=`pwd`
   fi
 fi
 
 if [ "$WITH_PLOTS" = "true" ] ; then
   cd ${PLOTDIR}
-  make || exit 1
+  ${MAKE} || exit 1
 fi
 
 cd ${UTILDIR}
-make || exit 1
+${MAKE} || exit 1
+
+cd ${CURDIR}/..
+if [ "$FC" = "flang" ] && [ -d flang_extra ] ; then
+  echo making flang extras
+  cd flang_extra
+  ${MAKE} || exit 1
+fi
 
 cd ${CURDIR}
 if [ -f obj/aed_external.o ] ; then
   /bin/rm obj/aed_external.o
 fi
-make AEDBENDIR=$DAEDBENDIR AEDDMODIR=$DAEDDMODIR || exit 1
+${MAKE} AEDBENDIR=$DAEDBENDIR AEDDMODIR=$DAEDDMODIR || exit 1
 if [ "${DAEDDEVDIR}" != "" -a -d ${DAEDDEVDIR} ] ; then
   echo now build plus version
   /bin/rm obj/aed_external.o
-  make glm+ AEDBENDIR=$DAEDBENDIR AEDDMODIR=$DAEDDMODIR AEDRIPDIR=$DAEDRIPDIR AEDDEVDIR=$DAEDDEVDIR || exit 1
+  ${MAKE} glm+ AEDBENDIR=$DAEDBENDIR AEDDMODIR=$DAEDDMODIR AEDRIPDIR=$DAEDRIPDIR AEDDEVDIR=$DAEDDEVDIR || exit 1
 fi
 
 
@@ -199,7 +230,7 @@ if [ "$OSTYPE" = "Linux" ] ; then
       sed -i "s/version=$VERSRUL/version=$VERSION/" debian/rules
     fi
 
-    fakeroot make -f debian/rules binary || exit 1
+    fakeroot ${MAKE} -f debian/rules binary || exit 1
 
     cd ..
